# docs/swagger-output.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "Weather Station API",
    "description": "JSON REST API for recording weather station data"
  },
  "servers": [
    {
      "url": "http://localhost:8081/"
    }
  ],
  "paths": {
...
    "/users/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log user in",
        "description": "",
        "parameters": [
          {
            "name": "LoginData",
            "in": "body",
            "description": "User login credentials",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "type": {
                  "type": "string",
                  "example": "object"
                },
                "properties": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "User email address"
                        }
                      }
                    },
                    "password": {
                      "type": "object",
                      "properties": {
                        "type": {
                          "type": "string",
                          "example": "string"
                        },
                        "description": {
                          "type": "string",
                          "example": "User password"
                        }
                      }
                    }
                  }
                },
                "example": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "test123@gmail.com"
                    },
                    "password": {
                      "type": "string",
                      "example": "abc123!"
                    }
                  }
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "email": {
                    "example": "any"
                  },
                  "password": {
                    "example": "any"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/logout": {
      "post": {
        "summary": "Log user out",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Log user out by forgetting authentication key",
          "content": {
            "application/json": {
              "schema": {
                "authenticationKey": "string"
              },
              "example": {
                "authenticationKey": "01493656-2b1b-48ba-b634-0b916b48466f"
              }
            }
          }
        }
      }
    },
    "/users": {
      "post": {
        "summary": "Create a new user",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Creating a new user account (singular)",
          "content": {
            "application/json": {
              "schema": {
                "email": "string",
                "password": "string",
                "role": "string",
                "firstName": "string",
                "lastName": "string"
              },
              "example": {
                "email": "test123@gmail.com",
                "password": "abc123!",
                "role": "admin",
                "firstName": "Bridgette",
                "lastName": "Collett"
              }
            }
          }
        }
      },
      "get": {
        "summary": "Get all users",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          }
        },
        "requestBody": {
          "description": "Get a list of all users"
        }
      },
      "patch": {
        "summary": "Update an existing user by ID",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Creating a new user account (singular)",
          "content": {
            "application/json": {
              "schema": {
                "id": "string",
                "email": "string",
                "password": "string",
                "role": "string",
                "firstName": "string",
                "lastName": "string",
                "authenticationKey": "array",
                "items": "string",
                "nullable": "true"
              },
              "example": {
                "id": "64587fad553b66691a4b8e9d",
                "email": "test123@gmail.com",
                "password": "abc123!",
                "role": "admin",
                "firstName": "Bridgette",
                "lastName": "Collett",
                "authenticationKey": "01493656-2b1b-48ba-b634-0b916b48466f"
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete user by ID",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Delete a specific (singular) user by their ID",
          "content": {
            "application/json": {
              "schema": {
                "id": "string"
              },
              "example": {
                "id": "64587fad553b66691a4b8e9d"
              }
            }
          }
        }
      }
    },
    "/user/{id}": {
      "get": {
        "summary": "Get user by ID",
        "description": "",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Get user and account details by searching by ID",
          "content": {
            "application/json": {
              "schema": {
                "id": "string"
              },
              "example": {
                "id": "64587fad553b66691a4b8e9d"
              }
            }
          }
        }
      }
    },
    "/users/update-many": {
      "patch": {
        "summary": "Update Many Users",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Update a batch of users by a date range",
          "content": {
            "application/json": {
              "schema": {
                "role": "string",
                "fromDate": "string",
                "toDate": "string"
              },
              "example": {
                "role": "admin",
                "fromDate": "2023-05-19",
                "toDate": "2023-05-21"
              }
            }
          }
        }
      }
    },
    "/users/delete-many": {
      "delete": {
        "summary": "Delete Many Users",
        "description": "",
        "parameters": [],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        },
        "requestBody": {
          "description": "Delete many specific (plural) users by their IDs",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "ids": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              },
              "example": {
                "ids": [
                  "64675b05e96736b4062c4eb7",
                  "64675af0e96736b4062c4eb5"
                ]
              }
            }
          }
        }
      }
    },
...
  }
}